pyproject.toml
[tool.poetry]
name = "books-project"
version = "0.1.0"
description = ""
authors = ["José Pedro Monteiro <zepedromonteiro@hotmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
django = "^5.1.5"
djangorestframework = "^3.15.2"
djangorestframework-simplejwt = "^5.4.0"
psycopg2-binary = "^2.9.10"
python-dotenv = "^1.0.1"


[tool.poetry.group.dev.dependencies]
yapf = "^0.43.0"
isort = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
